% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{all_estimates}
\alias{all_estimates}
\title{Extract RMST And Survival Probability Estimates}
\usage{
all_estimates(x)
}
\arguments{
\item{x}{An object of class "rmst" or "survprob".}
}
\value{
A data frame containing the estimates.
}
\description{
Extract RMST And Survival Probability Estimates
}
\examples{
\donttest{
surv <- survrct(Surv(time, status) ~ trt + age + sex + obstruct +
                   perfor + adhere + surg,
                target = "trt", data = colon, coarsen = 30, estimator = "tmle")
est <- rmst(surv, 105:111)
all_estimates(est)
}
}
\seealso{
\code{\link{rmst}} and \code{\link{survprob}} for creating \code{x}.
}
