% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{c19.ordinal}
\alias{c19.ordinal}
\title{Synthetic COVID-19 Data}
\format{
An object of class \code{data.frame} with 1500 rows and 13 columns.
}
\usage{
c19.ordinal
}
\description{
TODO
}
\keyword{datasets}
