% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{c19.tte}
\alias{c19.tte}
\title{Synthetic COVID-19 Data}
\format{
An object of class \code{data.frame} with 1500 rows and 14 columns.
}
\usage{
c19.tte
}
\description{
TODO
}
\keyword{datasets}
