% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{cdf}
\alias{cdf}
\title{Estimate the Cumulative Distribution Function}
\usage{
cdf(metadata)
}
\arguments{
\item{metadata}{An object of class "Ordinal" generated by a call to \code{ordinalrct()}.}
}
\value{
A list of class \code{cdf} containing the following components:

\item{estimator}{The estimation method used.}
\item{estimates}{The computed estimates including the
                      efficient influence function, standard errors,
                      and confidence intervals.}
}
\description{
Estimate the Cumulative Distribution Function
}
\seealso{
\code{\link{ordinalrct}} for creating \code{metadata}.

Other ordinalrct functions: 
\code{\link{log_or}()},
\code{\link{mannwhitney}()},
\code{\link{pmf}()}
}
\concept{ordinalrct functions}
