% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colon}
\alias{colon}
\title{Chemotherapy for Stage B/C colon cancer}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 614 rows and 15 columns.
}
\usage{
colon
}
\description{
Chemotherapy for Stage B/C colon cancer
}
\keyword{datasets}
