% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{log_or}
\alias{log_or}
\title{Estimate Average Log Odds Ratio}
\usage{
log_or(metadata)
}
\arguments{
\item{metadata}{An object of class "Survival" generated by a call to \code{metadata()}.}
}
\value{
A list of class \code{lor} containing the following components:

\item{estimator}{The estimation method used.}
\item{standard_error}{The computed estimates including the
                      efficient influence function, standard errors,
                      and confidence intervals.}
}
\description{
Estimate Average Log Odds Ratio
}
\seealso{
\code{\link{ordinalrct}} for creating \code{metadata}.
}
\concept{ordinalrct functions}
