% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{mannwhitney}
\alias{mannwhitney}
\title{Estimate the Mann-Whitney Estimand}
\usage{
mannwhitney(metadata)
}
\arguments{
\item{metadata}{An object of class "Ordinal" generated by a call to \code{ordinalrct()}.}
}
\value{
A list of class \code{mannwhit} containing the following components:

\item{estimator}{The estimation method used.}
\item{estimates}{The computed estimates including the
                      efficient influence function, standard errors,
                      and confidence intervals.}
}
\description{
Estimate the Mann-Whitney Estimand
}
\seealso{
\code{\link{ordinalrct}} for creating \code{metadata}.

Other ordinalrct functions: 
\code{\link{cdf}()},
\code{\link{log_or}()},
\code{\link{pmf}()}
}
\concept{ordinalrct functions}
