% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{metadata}
\alias{metadata}
\title{Create RCT Survival Metadata}
\usage{
metadata(
  data,
  trt,
  status,
  baseline,
  time,
  id,
  coarsen = 1,
  estimator = c("tmle", "aipw", "km"),
  lrnrs_trt = NULL,
  lrnrs_cens = NULL,
  lrnrs_hzrd = NULL
)
}
\arguments{
\item{lrnrs_hzrd}{}
}
\value{

}
\description{
Create RCT Survival Metadata
}
