% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{survprob}
\alias{survprob}
\title{Estimate Survival Probabilities}
\usage{
survprob(metadata, horizon = NULL)
}
\arguments{
\item{metadata}{A rctSurv Survival metadata object generated by a call to \code{metadata()}.}

\item{horizon}{The time horizon to evaluate at. If \code{NULL}, computes
survival probabilities at all time points.}
}
\value{

}
\description{
Estimate Survival Probabilities
}
