% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survrct-package.R
\docType{package}
\name{survrct-package}
\alias{survrct}
\alias{survrct-package}
\title{survrct: Efficient Estimators for Survival Analysis in RCTs Without Proportional Hazards}
\description{
Adjusted and unadjusted estimators for the restriced mean survival 
  time and survival probability in randomized controlled trials without the 
  proportional hazards assumptions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nt-williams/survrct}
  \item Report bugs at \url{https://github.com/nt-williams/survrct/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Williams \email{niw4001@med.cornell.edu} (\href{https://orcid.org/0000-0002-1378-4831}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Iván Díaz \email{ild2005@med.cornell.edu} (\href{https://orcid.org/0000-0001-9056-2047}{ORCID}) [copyright holder]
}

}
\keyword{internal}
