% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.cdf}
\alias{tidy.cdf}
\title{Tidy a(n) cdf object}
\usage{
\method{tidy}{cdf}(x, ...)
}
\arguments{
\item{x}{A `cdf` object.}

\item{...}{Unused, included for generic consistency only.}
}
\value{
A tidy [tibble::tibble()] summarizing CDF estimates.
}
\description{
Tidy a(n) cdf object
}
\examples{
rct <- ordinalrct(state_ordinal ~ A + age, A ~ 1, data = c19.ordinal)
tidy(cdf(rct))

}
