% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.lor}
\alias{tidy.lor}
\title{Tidy a(n) lor object}
\usage{
\method{tidy}{lor}(x, ...)
}
\arguments{
\item{x}{A `lor` object.}

\item{...}{Unused, included for generic consistency only.}
}
\value{
A tidy [tibble::tibble()] summarizing log odds estimates.
}
\description{
Tidy a(n) lor object
}
\examples{
rct <- ordinalrct(state_ordinal ~ A + age, A ~ 1, data = c19.ordinal)
tidy(log_or(rct))

}
