% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.mannwhit}
\alias{tidy.mannwhit}
\title{Tidy a(n) mannwhit object}
\usage{
\method{tidy}{mannwhit}(x, ...)
}
\arguments{
\item{x}{A `mannwhit` object.}

\item{...}{Unused, included for generic consistency only.}
}
\value{
A tidy [tibble::tibble()] summarizing the Mann-Whitney statistic estimates.
}
\description{
Tidy a(n) mannwhit object
}
\examples{
rct <- ordinalrct(state_ordinal ~ A + age, A ~ 1, data = c19.ordinal)
tidy(mannwhitney(rct))

}
