% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.pmf}
\alias{tidy.pmf}
\title{Tidy a(n) pmf object}
\usage{
\method{tidy}{pmf}(x, ...)
}
\arguments{
\item{x}{A `pmf` object.}

\item{...}{Unused, included for generic consistency only.}
}
\value{
A tidy [tibble::tibble()] summarizing PMF estimates.
}
\description{
Tidy a(n) pmf object
}
\examples{
rct <- ordinalrct(state_ordinal ~ A + age, A ~ 1, data = c19.ordinal)
tidy(pmf(rct))

}
