% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.rmst}
\alias{tidy.rmst}
\title{Tidy a(n) rmst object}
\usage{
\method{tidy}{rmst}(x, ...)
}
\arguments{
\item{x}{A `rmst` object.}

\item{...}{Unused, included for generic consistency only.}
}
\value{
A tidy [tibble::tibble()] summarizing restricted mean survival time estimates.
}
\description{
Tidy a(n) rmst object
}
\examples{
surv <- survrct(Surv(days, event) ~ A + age + sex + dyspnea + bmi,
                A ~ 1, data = c19.tte)
tidy(rmst(surv, 14))

}
