% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.survprob}
\alias{tidy.survprob}
\title{Tidy a(n) survprob object}
\usage{
\method{tidy}{survprob}(x, ...)
}
\arguments{
\item{x}{A `survprob` object.}

\item{...}{Unused, included for generic consistency only.}
}
\value{
A tidy [tibble::tibble()] summarizing survival probability estimates.
}
\description{
Tidy a(n) survprob object
}
\examples{
surv <- survrct(Surv(days, event) ~ A + age + sex + dyspnea + bmi,
                A ~ 1, data = c19.tte)
tidy(survprob(surv, 14))

}
