% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_fits}
\alias{get_fits}
\title{Return Nuisance Parameter Model Fits}
\usage{
get_fits(metadata)
}
\arguments{
\item{metadata}{An object of class "Survival" generated by a call to \code{metadata()}.}
}
\value{
A list containing the following components:

\item{Hazard}{The fit for the outcome model.}
\item{Censoring}{The fit for the censoring model.}
\item{Treatment}{The fit for the treatment model.}
}
\description{
Return Nuisance Parameter Model Fits
}
\examples{
\donttest{
surv <- survrct(Surv(time, status) ~ trt + age + sex + obstruct +
                   perfor + adhere + surg,
                target = "trt", data = colon, coarsen = 30, estimator = "tmle")
get_fits(surv)
}
}
\seealso{
\code{\link{survrct}} for creating \code{metadata}.
}
