% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{log_or}
\alias{log_or}
\title{Estimate Average Log Odds Ratio}
\usage{
log_or(metadata)
}
\arguments{
\item{metadata}{An object of class "Ordinal" generated by a call to \code{ordinalrct()}.}
}
\value{
A list of class \code{lor} containing the computed estimates including the
  efficient influence function, standard errors, and confidence intervals.
}
\description{
Estimate Average Log Odds Ratio
}
\examples{
rct <- ordinalrct(state_ordinal ~ A + age, A ~ 1, data = c19.ordinal)
log_or(rct)
}
\seealso{
\code{\link{ordinalrct}} for creating \code{metadata}.

Other ordinalrct functions: 
\code{\link{cdf}()},
\code{\link{mannwhitney}()},
\code{\link{pmf}()}
}
\concept{ordinalrct functions}
