% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{ordinalrct}
\alias{ordinalrct}
\title{Create RCT Ordinal Metadata}
\usage{
ordinalrct(
  outcome.formula,
  trt.formula,
  data,
  algo = c("glm", "lasso", "rf", "xgboost"),
  crossfit = TRUE
)
}
\arguments{
\item{outcome.formula}{An object of class "formula". The left hand side should be
the outcome variable. The right hand
side should contain the target variable of interest and baseline prognostic covariates.}

\item{trt.formula}{An object of class "formula". The left hand side should be the name
of the binary (coded as 0 and 1) treatment variable. The right hand side should specify the
names of variables used to estimate the propensity.}

\item{data}{A data frame containing the variables in the model.}

\item{algo}{Method to be used for fitting nuisance parameters. Automatically set to "glm" if
less than two covariates for adjustment are specified in \code{formula}.}

\item{crossfit}{Should the estimator be crossfit? Ignored if \code{algo} is \code{"glm"} or \code{"lasso"}}
}
\value{
An R6 object of class "Ordinal".
}
\description{
Compute metadata to be used for estimation of model-robust covariate adjusted
average log odds ratio, Mann-Whitney statistic, and counterfactual marginal CDF
and PMF in a two-arm randomized controlled trial using
targeted maximum likelihood estimation.
}
\examples{
ordinalrct(state_ordinal ~ A + age, A ~ 1, data = c19.ordinal)
}
\concept{ordinal functions}
