% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{ordinalrct}
\alias{ordinalrct}
\title{Create RCT Ordinal Metadata}
\usage{
ordinalrct(
  formula,
  target,
  data,
  estimator = c("tmle", "aipw"),
  algo = c("glm", "lasso", "rf")
)
}
\arguments{
\item{formula}{An object of class "formula". The left hand side should be
the outcome variable. The right hand
side should contain the target variable of interest and baseline prognostic covariates.}

\item{target}{The name of the target variable of interest; i.e., the randomized condition.}

\item{data}{A data frame containing the variables in the model.}

\item{estimator}{The method to be used for computing estimands. Options are "tmle"
for Targeted Maximum Likelihood Estimation (the default), "aipw" for Augmented IPW,
and "unadjusted" for the unadjusted estimator.}

\item{algo}{Method to be used for fitting nuisance parameters. Automatically set to "glm" if
less than two covariates for adjustment are specified in \code{formula}.}
}
\value{
An R6 object of class "Ordinal".
}
\description{
Create RCT ordinal metadata to be used for estimation of the
average log odds ratio, counterfactual marginal CDFs, and Mann-Whitney statistic.
}
\concept{ordinal functions}
