% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.rmst}
\alias{plot.rmst}
\title{Plot Restricted Mean Survival Time Estimates}
\usage{
\method{plot}{rmst}(x, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Plot Restricted Mean Survival Time Estimates
}
