% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.rmst}
\alias{plot.rmst}
\title{Plot Restricted Mean Survival Time Estimates}
\usage{
\method{plot}{rmst}(x, ...)
}
\arguments{
\item{x}{An object of class "rmst" produced by a call to \code{rmst()}}

\item{...}{Extra arguments to be passed to plot, not used.}
}
\value{
Object of class \code{ggplot} containing a patchwork step function plot of the
 treatment and control arm RMST point estimates and their differences across a sequence
 of time horizons.
}
\description{
Plot Restricted Mean Survival Time Estimates
}
\examples{
\donttest{
if (requireNamespace("survival", quietly = TRUE)) {
  library(survrct)
  veteran <- survival::veteran
  veteran$trt <- veteran$trt - 1
  veteran$celltype <- factor(veteran$celltype)
  surv <- survrct(Surv(time, status) ~ trt + celltype + karno + diagtime + age + prior,
                  target = "trt", data = veteran, coarsen = 7, estimator = "tmle")
  estm <- rmst(surv, 10:40)
  plot(estm)
  }
}
}
