% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.survprob}
\alias{plot.survprob}
\title{Plot Survival Probability Estimates}
\usage{
\method{plot}{survprob}(x, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Plot Survival Probability Estimates
}
