% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.survprob}
\alias{plot.survprob}
\title{Plot Survival Probability Estimates}
\usage{
\method{plot}{survprob}(x, ...)
}
\arguments{
\item{x}{An object of class "survprob" produced by a call to \code{survprob()}}

\item{...}{Extra arguments to be passed to plot, not used.}
}
\value{
Object of class \code{ggplot} containing a patchwork step function plot of the
 treatment and control arm survpval probability point estimates and their differences
 across a sequence of time horizons.
}
\description{
Plot Survival Probability Estimates
}
\examples{
\donttest{
surv <- survrct(Surv(time, status) ~ trt + age + sex + obstruct +
                   perfor + adhere + surg,
                target = "trt", data = colon, coarsen = 30, estimator = "tmle")
plot(survprob(surv, 105:111))
}
}
