% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{rmst}
\alias{rmst}
\title{Estimate Restricted Mean Survival Time}
\usage{
rmst(metadata, horizon = NULL)
}
\arguments{
\item{metadata}{A rctSurv Survival metadata object generated by a call to \code{metadata()}.}

\item{horizon}{The time horizon to evaluate at. If \code{NULL}, computes
RMST at all possible time horizons.}
}
\value{

}
\description{
Estimate Restricted Mean Survival Time
}
