% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{rmst}
\alias{rmst}
\title{Estimate Restricted Mean Survival Time}
\usage{
rmst(metadata, horizon = NULL)
}
\arguments{
\item{metadata}{An object of class "Survival" generated by a call to \code{survrct()}.}

\item{horizon}{The time horizon to evaluate at. If \code{NULL}, computes
RMST at all possible time horizons.}
}
\value{
A list of class \code{rmst} containing the following components:

\item{estimator}{The estimation method used.}
\item{horizon}{The time horizons RMST was evaluated at.}
\item{estimates}{The computed estimates for each time horizon including
                      the efficient influence function, standard errors, and
                      confidence intervals.}
}
\description{
Estimate Restricted Mean Survival Time
}
\examples{
\donttest{
surv <- survrct(Surv(time, status) ~ trt + age + sex + obstruct +
                   perfor + adhere + surg,
                target = "trt", data = colon, coarsen = 30, estimator = "tmle")
rmst(surv, 60)
}
}
\seealso{
\code{\link{survrct}} for creating \code{metadata}.

Other survrct functions: 
\code{\link{survprob}()},
\code{\link{survrct}()}
}
\concept{survrct functions}
