% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{survprob}
\alias{survprob}
\title{Estimate Survival Probabilities}
\usage{
survprob(metadata, horizon = NULL)
}
\arguments{
\item{metadata}{An object of class "Survival" generated by a call to \code{metadata()}.}

\item{horizon}{The time horizon to evaluate at. If \code{NULL}, computes
survival probabilities at all time points.}
}
\value{
A list of class \code{survprob} containing the following components:

\item{estimator}{The estimation method used.}
\item{horizon}{The time horizons the survival probability was evaluated at.}
\item{standard_error}{The computed estimates for each time horizon including the efficient influence function,
     standard errors, and confidence intervals.}
}
\description{
Estimate Survival Probabilities
}
\examples{
if (requireNamespace("survival", quietly = TRUE)) {
  veteran <- survival::veteran
  veteran$trt <- veteran$trt - 1
  veteran$celltype <- factor(veteran$celltype)
  surv <- survrct(Surv(time, status) ~ trt + celltype + karno + diagtime + age + prior,
                  target = "trt", data = veteran, coarsen = 7, estimator = "tmle")
  survprob(surv, horizon = 50)
}
}
\seealso{
\code{\link{survrct}} for creating \code{metadata}.

Other survrct functions: 
\code{\link{rmst}()},
\code{\link{survrct}()}
}
\concept{survrct functions}
