% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{survprob}
\alias{survprob}
\title{Estimate Survival Probabilities}
\usage{
survprob(metadata, horizon = NULL)
}
\arguments{
\item{metadata}{An object of class "Survival" generated by a call to \code{survrct()}.}

\item{horizon}{The time horizon to evaluate at. If \code{NULL}, computes
survival probabilities at all time points.}
}
\value{
A list of class \code{survprob} containing the following components:

\item{estimator}{The estimation method used.}
\item{horizon}{The time horizons the survival probability was evaluated at.}
\item{estimates}{The computed estimates for each time horizon including the
                      efficient influence function, standard errors,
                      and confidence intervals.}
}
\description{
Estimate Survival Probabilities
}
\examples{
surv <- survrct(Surv(days, event) ~ A + age + sex + dyspnea + bmi,
                A ~ 1, data = c19.tte)
survprob(surv, 14)
}
\seealso{
\code{\link{survrct}} for creating \code{metadata}.

Other survrct functions: 
\code{\link{rmst}()},
\code{\link{survrct}()}
}
\concept{survrct functions}
