% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{survprob}
\alias{survprob}
\title{Title}
\usage{
survprob(
  data,
  trt,
  status,
  baseline,
  time,
  id,
  horizon = NULL,
  coarsen = 1,
  estimator = c("tmle", "aipw", "ipw", "km"),
  learners_trt = NULL,
  learners_cens = NULL,
  learners_hazard = NULL
)
}
\arguments{
\item{learners_hazard}{}
}
\value{

}
\description{
Title
}
