% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{survrct}
\alias{survrct}
\title{Create RCT Survival Metadata}
\usage{
survrct(
  formula,
  target,
  data,
  coarsen,
  estimator,
  lrnrs_cens = NULL,
  lrnrs_hzrd = NULL
)
}
\arguments{
\item{lrnrs_hzrd}{}
}
\value{

}
\description{
Create RCT Survival Metadata
}
