% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{survrct}
\alias{survrct}
\title{Create RCT Survival Metadata}
\usage{
survrct(
  formula,
  target,
  data,
  coarsen = 1,
  estimator = c("tmle", "aipw"),
  algo = c("glm", "lasso", "rf")
)
}
\arguments{
\item{formula}{An object of class "formula". The left hand side should be
specified using \code{Surv(time, status)} where time is the name of the
variable containing follow-up time and status is the name of the indicator variable
with 0 = right censored and 1 = event at \code{time}. The right hand
side should contain the target variable of interest and baseline prognostic covariates.}

\item{target}{The name of the target variable of interest; i.e., the randomized condition.}

\item{data}{A data frame containing the variables in the model.}

\item{coarsen}{An optional integer that specifies if and how \code{time} should
be categorized into coarser intervals.}

\item{estimator}{The method to be used for computing estimands. Options are "tmle"
for Targeted Maximum Likelihood Estimation (the default) and "aipw" for Augmented IPW.}

\item{algo}{Method to be used for fitting nuisance parameters. Automatically set to "glm" if
less than two covariates for adjustment are specified in \code{formula}.}
}
\value{
An R6 object of class "Survival".
}
\description{
Create RCT survival metadata to be used for estimation of the
restricted mean survival time and survival probability.
}
\examples{
\donttest{
survrct(Surv(time, status) ~ trt + age + sex + obstruct + perfor + adhere + surg,
        target = "trt", data = colon, coarsen = 30, estimator = "tmle")
}
}
\seealso{
Other survrct functions: 
\code{\link{rmst}()},
\code{\link{survprob}()}
}
\concept{survrct functions}
