% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{survrct}
\alias{survrct}
\title{Create RCT Survival Metadata}
\usage{
survrct(
  outcome.formula,
  trt.formula,
  data,
  coarsen = 1,
  estimator = c("tmle", "aipw"),
  algo = c("glm", "lasso", "rf", "xgboost"),
  crossfit = TRUE
)
}
\arguments{
\item{outcome.formula}{An object of class "formula". The left hand side should be
specified using \code{Surv(time, status)} where time is the name of the
variable containing follow-up time and status is the name of the indicator variable
with 0 = right censored and 1 = event at \code{time}. The right hand
side should contain the target variable of interest and baseline prognostic covariates.}

\item{trt.formula}{An object of class "formula". The left hand side should be the name
of the binary (coded as 0 and 1) treatment variable. The right hand side should specify the
names of variables used to estimate the propensity.}

\item{data}{A data frame containing the variables in the model.}

\item{coarsen}{An optional integer that specifies if and how \code{time} should
be categorized into coarser intervals.}

\item{estimator}{The method to be used for computing estimands. Options are \code{"tmle"}
for Targeted Maximum Likelihood Estimation (the default) and \code{"aipw"} for Augmented IPW.}

\item{algo}{Method to be used for fitting the hazard and censoring nuisance parameters.
Automatically set to \code{"glm"} if less than two covariates for adjustment are
specified in \code{outcome.formula}. The propensity is always estimated using a GLM.}

\item{crossfit}{Should the estimator be crossfit? Ignored if \code{algo} is \code{"glm"} or \code{"lasso"}.}
}
\value{
An R6 object of class "Survival".
}
\description{
Compute metadata to be used for estimation of model-robust covariate adjusted
restricted mean survival time and survival probability in a two-arm randomized
controlled trial. Estimation may be done using targeted maximum likelihood estimation
or augmented inverse probability weighting.
}
\examples{
survrct(Surv(days, event) ~ A + age + sex + dyspnea + bmi,
        A ~ 1, data = c19.tte)
}
\seealso{
Other survrct functions: 
\code{\link{rmst}()},
\code{\link{survprob}()}
}
\concept{survrct functions}
